% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursion_functions.R
\name{backward_recursion}
\alias{backward_recursion}
\title{Backward Recursion Algorithm}
\usage{
backward_recursion(numObs, numStates, a = NULL, b = NULL, ct = NULL)
}
\description{
@description
A function used to calculate backward variables.
}
\details{
@details
This function is left to show the programming done for the Backward step in the Forward-Backward Algorithm
made for a Hidden Markov Model.

 @param numObs Number of Observations

 The total number of observations in your data set.

 @param numStates Number of States

 The total number of states planned for the Hidden Markov Model

 @param a State Transition Matrix

 The current state transition matrix. Input can be either a matrix or an array.

 @param b State Emission Matrix

 The current state emission matrix. Input can be either a matrix or an array.

 @param ct Normalizing Constant

 A vector containing constants normalizing `F_tilde`, the Forward Variables. This vector is typically
 found from running the Forward part of the Forward Algorithm.

 @returns A matrix of all your backward variables:
 * The rows correspond to the observations
 * The columns correspond to the states.

 @examples
 n = 100
 K = 3
 A = matrix(0,nrow = S, ncol = S)
 B = array(1, dim = c(n,K))
 fvars <- forward_recursion(numObs = n, numStates = K, a = A, b = B, ct = fvars[[2]])
 backward_recursion(numObs = n, numStates = K, )

 @export
}
