% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbaydesc.R
\docType{data}
\name{cbayrain}
\alias{cbayrain}
\title{Chesapeake Bay Rain data}
\format{
\subsection{\code{cbayrain}}{

A data frame with 1840 rows and 1928 columns:
\describe{
\item{Date}{2000-07-01 to 2019-09-30. Each row is a different day.}
\item{Locations no. 1-1927}{Each column is a different location}
}
}
}
\usage{
cbayrain
}
\description{
A subset of data containing the collected rainfall over several locations.
}
\keyword{datasets}
