% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_functions.R
\name{ELBO}
\alias{ELBO}
\title{ELBO Exponential Function}
\usage{
ELBO(
  numStates,
  numMix,
  numLoc,
  stateProb,
  mixProb,
  initProb,
  jtTransMat,
  ct,
  xi,
  alpha,
  zeta,
  gamma_shape,
  gamma_rate,
  obs,
  h
)
}
\arguments{
\item{numStates}{Number of States laid out in the model.}

\item{numMix}{Number of Mixtures laid out in the model.}

\item{numLoc}{Number of locations collected in the data.}

\item{stateProb}{Matrix of stationary probabilities.}

\item{mixProb}{Matrix of values used to weight the mixtures.}

\item{initProb}{Matrix of Initial Values for each state.}

\item{jtTransMat}{Transition Probability Matrix.}

\item{ct}{A matrix of values used to normalize the forward variables.}

\item{xi}{The Current Matrix for the Initial States.}

\item{alpha}{The Current Matrix for the Transitions.}

\item{zeta}{The Current Matrix for the Mixtures.}

\item{gamma_shape}{The Current matrix for the Shape Parameters}

\item{gamma_rate}{The Current Matrix for the Rate Parameters}

\item{obs}{The vector of observations.}

\item{h}{A matrix of values used in calculating the ELBO.}
}
\value{
A single value representing the ELBO for the given model parameters.
}
\description{
A function designed to calculate an approximation of the ELBO when using an exponential function to predict rainfall.
}
\details{
Since the ELBO is computationally difficult, or sometimes impossible, the function was designed to approximated it.
This function is used within the Variational Bayes EM loop to check for improvements.
}
