% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega_formulas.R
\name{exp_l_omega}
\alias{exp_l_omega}
\title{Omega Log-Gamma Function}
\usage{
exp_l_omega(gamh, delh, thetah, logbetah, con)
}
\arguments{
\item{gamh}{Value of the gamma hyperparameter of the GC2 distribution.}

\item{delh}{Value of the delta hyperparameter of the GC2 distribution.}

\item{thetah}{Value of the theta hyperparameter of the GC2 distribution.}

\item{logbetah}{Value of the log of beta hyperparameter of the GC2 distribution.}

\item{con}{The normalizing constant, typically found using the \code{omega_constant} function.}
}
\value{
A single value representing the expectation of the log-gamma of the shape parameter.
}
\description{
A function designed to find the expectation of the log-gamma of the shape parameter in a GC2 distribution.
}
\details{
In Variational Bayes, when dealing with a Gamma distribution, having both the shape and the rate parameter unknown
forces a new conjugate prior known as a GC2 distribution. The marginal of the rate parameter is another gamma, but
the marginal of the omega is a distribution that requires numerical integration. This function is made to find the
log-gamma of the shape parameter in a GC2 distribution.
}
