% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega_formulas.R
\name{exp_psi_omega}
\alias{exp_psi_omega}
\title{Omega Psi-Gamma Function}
\usage{
exp_psi_omega(gamh, delh, thetah, logbetah, con)
}
\arguments{
\item{gamh}{Value of the gamma hyperparameter of the GC2 distribution.}

\item{delh}{Value of the delta hyperparameter of the GC2 distribution.}

\item{thetah}{Value of the theta hyperparameter of the GC2 distribution.}

\item{logbetah}{Value of the log of beta hyperparameter of the GC2 distribution.}

\item{con}{The normalizing constant, typically found using the \code{omega_constant} function.}
}
\value{
A single value representing special expectation of the shape parameter.
}
\description{
A function designed to find a special expression of shape parameter in a GC2 distribution.
}
\details{
In Variational Bayes, when dealing with a Gamma distribution, having both the shape and the rate parameter unknown
forces a new conjugate prior known as a GC2 distribution. The marginal of the rate parameter is another gamma, but
the marginal of the omega is a distribution that requires numerical integration. The expression E(omega times Psi(omega * gammah))
appears when calculating the ELBO (in particular the KL Divergence of the GC2 distribution) and the expectation step in
an EM algorithm. Omegga is the shape parameter, gamma is the gamma hyperparameter of the GC2 distribution and Psi refers to the
derivative of the log of the gamma distribution.
}
