% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD2_Docu.R
\docType{data}
\name{Faux2}
\alias{Faux2}
\title{A Fake Data set from simulating a HMM}
\format{
A list containing rainfall amounts.
}
\source{
Generated using a fake HMM mode.
}
\usage{
Faux2
}
\description{
A dataset containing the rainfall amounts and the state associated with said rainfall amount coming from 10 locations over 20 years.
}
\keyword{datasets}
