% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmmGen.R
\name{generate.hmm}
\alias{generate.hmm}
\title{HMM Simulator}
\usage{
generate.hmm(
  D,
  Y = 1,
  L = 1,
  S,
  M = 2,
  initDist,
  tmat,
  zeta,
  dist = c("exp", "gamma"),
  lambda,
  omega
)
}
\description{
Creates synthetic data based on which distribution you choose.
}
