% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega_formulas.R
\name{omega_constant}
\alias{omega_constant}
\title{Omega Normalization Function}
\usage{
omega_constant(gamh, delh, thetah, logbetah)
}
\arguments{
\item{gamh}{Value of the gamma hyperparameter of the GC2 distribution.}

\item{delh}{Value of the delta hyperparameter of the GC2 distribution.}

\item{thetah}{Value of the theta hyperparameter of the GC2 distribution.}

\item{logbetah}{Value of the log of beta hyperparameter of the GC2 distribution.}
}
\value{
A single value representing the normalizing constant. Note that, to use in a function,
it would need to be 1/result.
}
\description{
A function designed to find the normalizing constant of the marginal distribution of a shape parameter in
a GC2 distribution.
}
\details{
In Variational Bayes, when dealing with a Gamma distribution, having both the shape and the rate parameter unknown
forces a new conjugate prior known as a GC2 distribution. The marginal of the rate parameter is another gamma, but
the marginal of the omega is a distribution that requires numerical integration. This function is made to find the
constant that normalizes this distribution.
}
