% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_param_function.R
\name{post_param.gam}
\alias{post_param.gam}
\title{Gamma Posterior Parameter Function}
\usage{
post_param.gam(
  numStates,
  numMix,
  gamma.post,
  delta.post,
  theta.post,
  exp_alpha,
  zeta,
  alpha,
  xi
)
}
\arguments{
\item{numStates}{Number of States

The total number of states planned for the Hidden Markov Model.}

\item{numMix}{Number of Mixtures

The total number of mixtures planned for the Hidden Markov Model.}

\item{gamma.post}{The posterior matrix for the Gamma Parameters

The posterior matrix of the shape hyperparameters for each GC2 prior.}

\item{delta.post}{The Posterior Matrix for the Delta Parameters

The posterior matrix of the rate hyperparameters for each GC2 prior.}

\item{theta.post}{The Posterior Matrix for the Theta Parameters

The posterior matrix of the rate hyperparameters for each GC2 prior.}

\item{exp_alpha}{The expectation of the unknown shape parameter.

The assumption in using a GC2 distribution is that both the shape and rate of a Gamma distribution
are unknown. Thus, quantities like the expectation of the shape parameter need to be calculated.
The commands \code{omega_constant} and \code{exp_omega} can be used to calculate these values.}

\item{zeta}{The Posterior Matrix for the Mixtures

The posterior matrix of the hyperparameters for the Dirichlet distribution used to describe
the mixture probabilities at each location.}

\item{alpha}{The Posterior Matrix for the Transitions.

The posterior matrix of the hyperparameters for the Dirichlet distribution used to describe
the transition probabilities from state to state.}

\item{xi}{The Posterior Matrix for the Initial States.

The posterior matrix of the hyperparameters for the Dirichlet distribution used to describe
the initial probabilities each state.}
}
\value{
A list holding four objects:
\itemize{
\item \code{InitDist} : A vector containing the posterior distribution of \code{pi_1}, the initial state probabilities.
\item \code{TransMat} : A matrix containing the posterior distribution of \code{A}, the transition state probabilities.
\item \code{MixProb} : A matrix containing the posterior distribution of \code{C}, the cluster mixture probabilities.
\item \code{RainRate} : A matrix containing the posterior distribution of \code{lambda}, the parameter of the exponential distribution
for rainfall at any given location.
}
}
\description{
A function used to put the posterior distributions of important parameters into a single list.
}
\details{
A function used to convert different outputs from an EM algorithm step into a single
consolidated list, while normalizing any variables that need to be normalized. Used at the very end of
an EM Algorithm step for a Variational Bayes Hidden Markov Model. This function is particular to using a
Gamma distribution to predict rainfall.
}
