% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_param_function.R
\name{post_param}
\alias{post_param}
\title{Posterior Parameter Function}
\usage{
post_param(numStates, numMix, gamma.post, delta.post, zeta, alpha, xi)
}
\arguments{
\item{numStates}{Number of States

The total number of states planned for the Hidden Markov Model.}

\item{numMix}{Number of Mixtures

The total number of mixtures planned for the Hidden Markov Model.}

\item{gamma.post}{The posterior matrix for the Shape Parameters

The posterior matrix of the shape hyperparameters for each Gamma mixture component.}

\item{delta.post}{The Posterior Matrix for the Rate Parameters

The posterior matrix of the rate hyperparameters for each Gamma mixture component.}

\item{zeta}{The Posterior Matrix for the Mixtures

The posterior matrix of the hyperparameters for the Dirichlet distribution used to describe
the mixture probabilities at each location.}

\item{alpha}{The Posterior Matrix for the Transitions.

The posterior matrix of the hyperparameters for the Dirichlet distribution used to describe
the transition probabilities from state to state.}

\item{xi}{The Posterior Matrix for the Initial States.

The posterior matrix of the hyperparameters for the Dirichlet distribution used to describe
the initial probabilities each state.}
}
\value{
A list holding four objects:
\itemize{
\item \code{InitDist} : A vector containing the posterior distribution of \code{pi_1}, the initial state probabilities.
\item \code{TransMat} : A matrix containing the posterior distribution of \code{A}, the transition state probabilities.
\item \code{MixProb} : A matrix containing the posterior distribution of \code{C}, the cluster mixture probabilities.
\item \code{RainRate} : A matrix containing the posterior distribution of \code{lambda}, the parameter of the exponential distribution
for rainfall at any given location.
}
}
\description{
A function used to put the posterior distributions of important parameters into a single list.
}
\details{
A function used to convert different datasets from an EM algorithm step into a single
consolidated list, while normalizing any variables that need to be normalized. Used at the very end of
an EM Algorithm step for a Variational Bayes Hidden Markov Model. This function is particular to using
Exponential Priors for each mixture.
}
\examples{
K = 2
M = 2
gamma = c(1,2)
delta = c(1,4)
zeta = matrix(c(1,1,1,1), nrow = K, ncol = M)
alpha = matrix(c(1,0,0,1), nrow = K, ncol = M)
xi = c(1,1,1,1)
post_var <- post_param(K, M, gamma, delta, zeta, alpha, xi)
print(post_var$RainRate)

}
