% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBE_function.R
\name{VBE.exp}
\alias{VBE.exp}
\title{Variational Bayes Expectation Formula}
\usage{
VBE.exp(
  numDays,
  numStates,
  numYears,
  numLoc,
  numMix,
  xi,
  alpha,
  zeta,
  gamma_shape,
  gamma_rate,
  var,
  obs
)
}
\arguments{
\item{numDays}{Number of Days per Year.}

\item{numStates}{Number of States laid out in the model.}

\item{numYears}{Number of Years collected.}

\item{numLoc}{Number of locations data recorded in the data.}

\item{numMix}{Number of mixtures predetermined by user.}

\item{xi}{The Current Matrix for the Initial States.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the initial probabilities each state.}

\item{alpha}{The Current Matrix for the Transitions.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the transition probabilities from state to state.}

\item{zeta}{The Current Matrix for the Mixtures.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the mixture probabilities at each location.}

\item{var}{A list of matrices and vectors objects:
\itemize{
\item \code{del_y0}: Rain indicator function. Typically an \code{ifelse} function.
\item \code{a_jk}: Posterior state kernel, a matrix filled with 0's with dimensions: \code{numStates} by \code{numStates}.
\item \code{b_tj}: Posterior emission kernel, an array filled with 1's with dimensions: \code{numDays} by \code{numStates} by \code{numYears}.
\item \code{b_tjl}: An augmented \code{b_tj} array, filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numYears} by \code{numLoc}.
\item \code{a_1j}: Posterior initial probability kernel, a matrix with dimensions: \code{numYears} by \code{numStates}.
\item \code{ct}: A matrix filled with 0's with dimensions: \code{numDays} by \code{numYears}.
\item \code{fvar}: An array filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numYears}.
\item \code{bvar}: An array filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numYears}.
\item \code{b_star}: An array filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numMix}.
\item \code{q_tj}: Posterior probability of stationary distribution, an array filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numYears}.
\item \code{q_tjml}: An augmented \code{q_tjml} array, filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numMix} by \code{numStates} by \code{numLoc}.
\item \code{q_jk}: Posterior joint transition probability matrix, an array filled with 0's with dimensions: \code{numStates} by \code{numStates} by \code{numDays} - 1 by \code{numYears}.
\item \code{q_1j}: Posterior initial probability, a matrix filled with 0's with dimensions: \code{numYears} by \code{numStates}.
}}

\item{obs}{The vector of observations.

The vector containing the y-values (typically the precipitation amounts) from the data.}

\item{gamma.post}{The Current matrix for the Shape Parameters.

The current matrix of the shape hyperparameters for each Gamma mixture component.}

\item{delta.post}{The Current Matrix for the Rate Parameters.

The current matrix of the rate hyperparameters for each Gamma mixture component.}
}
\value{
A list of objects built from empty objects in \code{var}:
\itemize{
\item \code{a_jk}: The posterior state kernel.
\item \code{b_tj}: Posterior emission kernel.
\item \code{c_t} : A series of constants used to normalize the forward variables.
\item \code{q_1j}: Posterior initial probability.
\item \code{q_tj}: Posterior probability of stationary distribution.
\item \code{q_tjml}: An augmented \code{q_tjml} array.
\item \code{q_jk}: Posterior joint transition probability matrix.
}
}
\description{
A function dedicated to the expectation step in a Variational Bayes EM algorithm
}
\details{
Used as part of the Variational Bayes EM Process for a Hidden Markov Model designed
for rain data.
}
