% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBEG_function.R
\name{VBE.gam}
\alias{VBE.gam}
\title{Variational Bayes Expectation Gamma Formula}
\usage{
VBE.gam(
  numDays,
  numStates,
  numYears,
  numLoc,
  numMix,
  xi,
  alpha,
  zeta,
  gamma,
  delta,
  logbeta,
  theta,
  exp_omega,
  exp_psi_omega,
  exp_lomega,
  var,
  obs
)
}
\arguments{
\item{numDays}{Number of Days per Year.}

\item{numStates}{Number of States laid out in the model.}

\item{numYears}{Number of Years collected.}

\item{numLoc}{Number of locations data recorded in the data.}

\item{numMix}{Number of mixtures predetermined by user.}

\item{xi}{The Current Matrix for the Initial States.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the initial probabilities each state.}

\item{alpha}{The Current Matrix for the Transitions.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the transition probabilities from state to state.}

\item{zeta}{The Current Matrix for the Mixtures.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the mixture probabilities at each location.}

\item{gamma}{The Current matrix for the gamma hyperparameters of the GC2 distribution.}

\item{delta}{The Current Matrix for the delta hyperparameters of the GC2 distribution.}

\item{logbeta}{The Current Matrix for the log of the beta hyperparameters of the GC2 distribution.}

\item{exp_omega}{A matrix of the expectation of the shape parameter.

The command \code{omega_constant} and \code{exp_omega} can be used to calculate these values.}

\item{exp_psi_omega}{A matrix of the expectation a special expectation of the shape parameter.

The command \code{omega_constant} and \code{exp_psi_omega} can be used to calculate these values.}

\item{exp_lomega}{A matrix of the expectation of the log-gamma of the shape parameter.

The command \code{omega_constant} and \code{exp_l_omega} can be used to calculate these values.}

\item{var}{A list of matrices and vectors objects:
\itemize{
\item \code{del_y0}: Rain indicator function. Typically an \code{ifelse} function.
\item \code{a_jk}: Posterior state kernel, a matrix filled with 0's with dimensions: \code{numStates} by \code{numStates}.
\item \code{b_tj}: Posterior emission kernel, an array filled with 1's with dimensions: \code{numDays} by \code{numStates} by \code{numYears}.
\item \code{b_tjl}: An augmented \code{b_tj} array, filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numYears} by \code{numLoc}.
\item \code{a_1j}: Posterior initial probability kernel, a matrix with dimensions: \code{numYears} by \code{numStates}.
\item \code{ct}: A matrix filled with 0's with dimensions: \code{numDays} by \code{numYears}.
\item \code{fvar}: An array filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numYears}.
\item \code{bvar}: An array filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numYears}.
\item \code{b_star}: An array filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numMix}.
\item \code{q_tj}: Posterior probability of stationary distribution, an array filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numYears}.
\item \code{q_tjml}: An augmented \code{q_tjml} array, filled with 0's with dimensions: \code{numDays} by \code{numStates} by \code{numMix} by \code{numStates} by \code{numLoc}.
\item \code{q_jk}: Posterior joint transition probability matrix, an array filled with 0's with dimensions: \code{numStates} by \code{numStates} by \code{numDays} - 1 by \code{numYears}.
\item \code{q_1j}: Posterior initial probability, a matrix filled with 0's with dimensions: \code{numYears} by \code{numStates}.
}}

\item{obs}{The vector of observations.

The vector containing the y-values (typically the precipitation amounts) from the data.}
}
\value{
A list of objects built from empty objects in \code{var}:
\itemize{
\item \code{a_jk}: The posterior state kernel.
\item \code{b_tj}: Posterior emission kernel.
\item \code{c_t} : A series of constants used to normalize the forward variables.
\item \code{q_1j}: Posterior initial probability.
\item \code{q_tj}: Posterior probability of stationary distribution.
\item \code{q_tjml}: An augmented \code{q_tjml} array.
\item \code{q_jk}: Posterior joint transition probability matrix.
}
}
\description{
A function dedicated to the expectation step in a Variational Bayes EM algorithm
}
\details{
Used as part of the Variational Bayes EM Process for a Hidden Markov Model designed
for rain data. It is assumed the rainfall is calculated using a Gamma Distribution.
}
