% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBEM_function.R
\name{VBEM}
\alias{VBEM}
\title{Variational Bayes EM Function}
\usage{
VBEM(
  D,
  S,
  Y,
  L,
  M,
  xi,
  alpha,
  zeta,
  gamma_shape,
  gamma_rate,
  obs,
  maxiter = 1000
)
}
\arguments{
\item{D}{Number of Days per Year}

\item{S}{Number of States laid out in the model}

\item{Y}{Number of Years collected}

\item{L}{Number of locations data recorded in the data}

\item{M}{Number of mixtures predetermined by user}

\item{xi}{The Initial Matrix for the Initial States.

The starting matrix of the hyperparameters for the Dirichlet distribution used to describe
the initial probabilities each state.}

\item{alpha}{The Initial Matrix for the Transitions.

The starting matrix of the hyperparameters for the Dirichlet distribution used to describe
the transition probabilities from state to state.}

\item{zeta}{The Initial Matrix for the Mixtures

The starting matrix of the hyperparameters for the Dirichlet distribution used to describe
the mixture probabilities at each location.}

\item{gamma_shape}{The Initial matrix for the Shape Parameters

The starting matrix of the shape hyperparameters for each Gamma mixture component.}

\item{gamma_rate}{The Initial Matrix for the Rate Parameters

The state matrix of the rate hyperparameters for each Gamma mixture component.}

\item{obs}{The vector of observations.

The vector containing the y-values (typically the precipitation amounts) from the data.}

\item{maxiter}{The maximum number of iterations the algorithm will do before stopping.

The algoritm will stop earlier than this number if the amount of improvement is smaller than a tolerance
 of 10^-9. The default value is 1000.}
}
\value{
A list of objects used to describe the model:
 * `priors`: A list containing all the starting matrices and probabilities.
 * `posteriors`: A list containing all the posterior matrices and probabilities.
 * `ELBO`: A vector tracking the ELBO as the model progresses (with spot `i` corresponding to iteration `i`)
 * `DIC`: A vector tracking the DIC as the model progresses (with spot `i` corresponding to iteration `i`)

 @export
}
\description{
A function dedicated to running the Variational Bayes EM algorithm.
}
\details{
Used to run a simple model with a low number of observations, locations and a fixed number of states and mixtures.
}
