% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBEmpfr_function.R
\name{VBEmpfr.exp}
\alias{VBEmpfr.exp}
\title{High Precision Variational Bayes Expectation Formula}
\usage{
VBEmpfr.exp(
  numDays,
  numStates,
  numLoc,
  numMix,
  xi,
  alpha,
  zeta,
  gamma_shape,
  gamma_rate,
  var,
  obs
)
}
\arguments{
\item{numDays}{Number of Days per Year.}

\item{numStates}{Number of States laid out in the model.}

\item{numLoc}{Number of locations data recorded in the data.}

\item{numMix}{Number of mixtures predetermined by user.}

\item{xi}{The Current Matrix for the Initial States.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
 the initial probabilities each state.}

\item{alpha}{The Current Matrix for the Transitions.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
 the transition probabilities from state to state.}

\item{zeta}{The Current Matrix for the Mixtures.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
 the mixture probabilities at each location.}

\item{var}{A list of matrices and vectors objects:
* `del_y0`: Rain indicator function. Typically an `ifelse` function.
* `a_jk`: Posterior state kernel, a matrix filled with 0's with dimensions: `numStates` by `numStates`.
* `b_tj`: Posterior emission kernel, an array filled with 1's with dimensions: `numDays` by `numStates` by `numYears`.
* `b_tjl`: An augmented `b_tj` array, filled with 0's with dimensions: `numDays` by `numStates` by `numYears` by `numLoc`.
* `a_1j`: Posterior initial probability kernel, a matrix with dimensions: `numYears` by `numStates`.
* `ct`: A matrix filled with 0's with dimensions: `numDays` by `numYears`.
* `fvar`: An array filled with 0's with dimensions: `numDays` by `numStates` by `numYears`.
* `bvar`: An array filled with 0's with dimensions: `numDays` by `numStates` by `numYears`.
* `b_star`: An array filled with 0's with dimensions: `numDays` by `numStates` by `numMix`.
* `q_tj`: Posterior probability of stationary distribution, an array filled with 0's with dimensions: `numDays` by `numStates` by `numYears`.
* `q_tjml`: An augmented `q_tjml` array, filled with 0's with dimensions: `numDays` by `numStates` by `numMix` by `numStates` by `numLoc`.
* `q_jk`: Posterior joint transition probability matrix, an array filled with 0's with dimensions: `numStates` by `numStates` by `numDays` - 1 by `numYears`.
* `q_1j`: Posterior initial probability, a matrix filled with 0's with dimensions: `numYears` by `numStates`.}

\item{obs}{The vector of observations.

The vector containing the y-values (typically the precipitation amounts) from the data.}

\item{numYears}{Number of Years collected.}

\item{gamma.post}{The Current matrix for the Shape Parameters.

The current matrix of the shape hyperparameters for each Gamma mixture component.}

\item{delta.post}{The Current Matrix for the Rate Parameters.

The current matrix of the rate hyperparameters for each Gamma mixture component.}
}
\value{
A list of objects built from empty objects in `var`:
 * `a_jk`: The posterior state kernel.
 * `b_tj`: Posterior emission kernel.
 * `c_t` : A series of constants used to normalize the forward variables.
 * `q_1j`: Posterior initial probability.
 * `q_tj`: Posterior probability of stationary distribution.
 * `q_tjml`: An augmented `q_tjml` array.
 * `q_jk`: Posterior joint transition probability matrix.
}
\description{
A function dedicated to the expectation step in a Variational Bayes EM algorithm with higher precision.
}
\details{
Used as part of the Variational Bayes EM Process for a Hidden Markov Model designed
 for rain data.
}
