% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Randomizer_Function.R
\name{Data_Randomizer}
\alias{Data_Randomizer}
\title{Data Randomizer Function}
\usage{
Data_Randomizer(numLoc, obs, mix.sample = F)
}
\arguments{
\item{numLoc}{The number of locations.}

\item{obs}{The dataset.}

\item{mix.samples}{The indicator used to decide the data should come from different years.

Initial set to FALSE.}
}
\value{
A list of objects:
\itemize{
\item \code{y_sample} : The resulting dataset with dimensions days by locations.
\item \code{del_y0_sample} : The indicator data set that uses 0 if no rain, 1 if there is any rain amount.
}
}
\description{
A function used to help sample during Stochastic VI.
}
\details{
In Stochastic VI, we split the data length N into D*Y, days and years. There are then two options on how to pick
data. Either, one can pick from the same year (all D data points come from the same year) or the data points
can come from all different years (D data points pick from different years).
}
