% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_functions.R
\name{ELBO.gam}
\alias{ELBO.gam}
\title{ELBO Gamma Function}
\usage{
ELBO.gam(
  numStates,
  numMix,
  numLoc,
  stateProb,
  mixProb,
  initProb,
  jtTransMat,
  ct,
  xi,
  alpha,
  zeta,
  gamma_hyper,
  delta_hyper,
  theta_hyper,
  logbetaprior,
  logbetapost,
  exp_omega,
  exp_psi_omega,
  exp_lomega,
  obs,
  h
)
}
\arguments{
\item{numStates}{Number of States laid out in the model.}

\item{numMix}{Number of Mixtures laid out in the model.}

\item{numLoc}{Number of locations collected in the data.}

\item{stateProb}{Matrix of stationary probabilities.}

\item{mixProb}{Matrix of values used to weight the mixtures.}

\item{initProb}{Matrix of Initial Values for each state.}

\item{jtTransMat}{Transition Probability Matrix.}

\item{ct}{A matrix of values used to normalize the forward variables.}

\item{xi}{The Current Matrix for the Initial States.}

\item{alpha}{The Current Matrix for the Transitions.}

\item{zeta}{The Current Matrix for the Mixtures.}

\item{gamma_hyper}{Value of the gamma hyperparameter of the GC2 distribution.}

\item{delta_hyper}{Value of the delta hyperparameter of the GC2 distribution.}

\item{theta_hyper}{Value of the theta hyperparameter of the GC2 distribution.}

\item{logbetaprior}{Value of the log of beta hyperparameter of the GC2 distribution before updating.}

\item{logbetapost}{Value of the log of beta hyperparameter of the GC2 distribution before after.}

\item{exp_omega}{A matrix of the expectation of the shape parameter.

The commands \code{omega_constant} and \code{exp_omega} can be used to calculate these values.}

\item{exp_psi_omega}{A matrix of the expectation a special expectation of the shape parameter.

The commands\code{omega_constant} and \code{exp_psi_omega} can be used to calculate these values.}

\item{exp_lomega}{A matrix of the expectation of the log-gamma of the shape parameter.

The commands \code{omega_constant} and \code{exp_l_omega} can be used to calculate these values.}

\item{obs}{The vector of observations.}

\item{h}{A matrix of values used in calculating the ELBO.

This value can be found as an output from the maximization in the EM algorithm (\code{VBM.gam}), or calculated manually
as it is the gamma_hyper*log(delta_hyper) minus the log-gamma(gamma_hyper).}
}
\value{
A single value representing the ELBO for the given model parameters.
}
\description{
A function designed to calculate an approximation of the ELBO when using a gamma function to describe rainfall.
}
\details{
Since the ELBO is computationally difficult, or sometimes impossible, the function was designed to approximated it.
This function is used within the Variational Bayes EM loop to check for improvements.
}
