% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FD1_Docu.R
\docType{data}
\name{Faux1}
\alias{Faux1}
\title{A Fake Data set from simulating a HMM}
\format{
A list containing:
\describe{
\item{rainfall}{The fake rainfall amount recorded over 1800 days at three different locations. Contains several
zero amounts.}
\item{hidden}{The hidden state associated with each rainfall calculation.}
\item{init_tmat}{The transition matrix used in generating the data.}
\item{init_pi}{The starting state probability matrix used in generating the data.}
\item{init_lambda}{The starting lambdas used in generating the rainfall amounts.}
\item{init_zeta}{The starting probabilities for the mixtures used in generating the rainfall amounts.}
}
}
\source{
Generated using a fake HMM mode.
}
\usage{
Faux1
}
\description{
A dataset containing the rainfall amounts and the state associated with said rainfall amount. The data simulates 1800
days, 3 mixtures, 3 hidden states, and 3 locations.
}
\keyword{datasets}
