% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitVBEM.R
\name{fit.VBEM}
\alias{fit.VBEM}
\title{Variational Bayes EM Algorithm for Rainfall}
\usage{
fit.VBEM(
  data,
  L,
  S,
  M = 2,
  D,
  Y = 1,
  dist = c("exp", "gamma", "stochexp"),
  hypers,
  mix.samples = F,
  maxiter = 1000
)
}
\arguments{
\item{data}{The recorded rainfall amounts. The amount must be in a matrix/array with dimensions L by T.}

\item{L}{Number of locations data recorded in the data.}

\item{S}{Number of States to put in the model.}

\item{M}{Number of mixtures predetermined by user.

Default value is set to 2, 1 mixture being the 0 rainfall mixture, the other being a single distribution of a
specified type.}

\item{D}{Number of Days per Year.}

\item{Y}{Number of Years collected.

Default value is set to 1.}

\item{dist}{The preferred distribution to describe rainfall.

As of now, there are only three options, "exp" for an exponential distribution, "stoch.exp" for a stochastic variational bayes of
an exponential distribution and "gamma" for a gamma distribution.}

\item{hypers}{A list containing all necessary hyperparameters in array form.

For any distribution choice, one must provide hyperparameters for xi, alpha and zeta. For exponential distributions,
the gamma shape and rate must be provided. For gamma distributions, four hyperparameters must all be provided (gamma, delta, theta, logbeta).
See "" for explanations. Please provide the hyperparameters in a list.}

\item{mix.samples}{A boolean that is only used if "stoch.exp" is chosen.

In Stochastic Variational Bayes, rather than using all data points, the data is broken into years, each with an equal
number of days. Using an exchangeability assumption, the model then only runs the model on a subset of the data. Keeping
mix.samples as \code{FALSE} means the data used is all from one year, changing it to \code{TRUE} means that for each day, the model
will randomly pick a year to sample from.}

\item{maxiter}{The maximum number of iterations the algorithm will do before stopping.

The default value is 1000.}
}
\value{
A list of objects used to describe the model:
\itemize{
\item \code{posteriors}: A list containing all the posterior matrices and probabilities.
\item \code{ELBO}: A vector tracking the ELBO as the model progresses (with spot \code{i} corresponding to iteration \code{i})
\item \code{DIC}: A vector tracking the DIC as the model progresses (with spot \code{i} corresponding to iteration \code{i})
\item \code{iternum}: An integer denoting how many iterations took place.
}
}
\description{
A function dedicated to running the Variational Bayes EM algorithm with several different types of models.
}
