% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursion_functions.R
\name{forward_recursion}
\alias{forward_recursion}
\title{Forward Recursion Algorithm}
\usage{
forward_recursion(
  numObs,
  numStates,
  initDist = c(1, rep(0, K - 1)),
  a = NULL,
  b = NULL
)
}
\description{
@description
}
\details{
A function used to calculate forward variables.

@details

This function is left to show the programming done for the Forward step in the Forward-Backward Algorithm
made for a Hidden Markov Model.

@param numObs Number of Observations

The total number of observations in your data set.

@param numStates Number of States

The total number of states planned for the Hidden Markov Model

@param initDist Initial Distribution

The initial probability kernel. If no value is given, it will automatically give
the first state a probability of 1, and all states a probability of 0.

@param a State Transition Matrix

The current state transition matrix. Input can be either a matrix or an array.

@param b State Emission Matrix

The current state emission matrix. Input can be either a matrix or an array.

@returns A list holding two objects:
\itemize{
\item \code{fvar_tilde} : A matrix containing all your Forward Variables. Each row corresponds to each observation
and each column corresponds to each state.
\item \code{ct} : A vector containing the constant used to normalize \code{F_tilde} at each point in time \code{t}.
}

@examples
n = 100
K = 3
A = matrix(0,nrow = S, ncol = S)
B = array(1, dim = c(n,K))
fvars <- forward_recursion(numObs = n, numStates = K, a = A, b = B)
print(fvars$fvar_tilde)
print(fvars$ct)

@export
}
