% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_functions.R
\name{StoDIC}
\alias{StoDIC}
\title{DIC Function}
\usage{
StoDIC(
  numStates,
  numMix,
  numLoc,
  stateProb,
  mixProb,
  initProb,
  jtTransMat,
  ct,
  xi,
  alpha,
  zeta,
  gamma_shape,
  gamma_rate
)
}
\arguments{
\item{numStates}{Number of States laid out in the model.}

\item{numMix}{Number of Mixtures laid out in the model.}

\item{numLoc}{Number of locations collected in the data.}

\item{stateProb}{Matrix of stationary probabilities.}

\item{mixProb}{Matrix of values used to weight the mixtures.}

\item{initProb}{Matrix of Initial Values for each state.}

\item{jtTransMat}{Transition Probability Matrix.}

\item{ct}{A matrix of values used to normalize the forward variables.}

\item{xi}{The Current Matrix for the Initial States.}

\item{alpha}{The Current Matrix for the Transitions.}

\item{zeta}{The Current Matrix for the Mixtures.}

\item{gamma_shape}{The Current matrix for the Shape Parameters}

\item{gamma_rate}{The Current Matrix for the Rate Parameters}
}
\value{
A list of two objects:
*\code{DIC}: A single value representing the DIC for the given model parameters.
*\code{pd}: A value used in calculating the DIC.
}
\description{
A function designed to calculate an approximation of the DIC.
}
\details{
Since the DIC is computationally difficult, or sometimes impossible, the function was designed to approximated it.
This function is used within the Variational Bayes EM loop to check for improvements, with SVI in mind.
}
