% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochasticVBEM_Function.R
\name{StoVBEM}
\alias{StoVBEM}
\title{Stochastic Variational Bayes Expectation Formula}
\usage{
StoVBEM(
  D,
  S,
  Y,
  L,
  M,
  xi,
  alpha,
  zeta,
  gamma_shape,
  gamma_rate,
  obs,
  mix.samples = F,
  maxiter = 1000
)
}
\arguments{
\item{D}{Number of Days per Year.}

\item{S}{Number of States laid out in the model.}

\item{Y}{Number of Years collected.}

\item{L}{Number of locations data recorded in the data.}

\item{M}{Number of mixtures predetermined by user.}

\item{xi}{The Initial Matrix for the Initial States.

The initial matrix of the hyperparameters for the Dirichlet distribution used to describe
 the initial probabilities each state.}

\item{alpha}{The Initial Matrix for the Transitions.

The initial matrix of the hyperparameters for the Dirichlet distribution used to describe
 the transition probabilities from state to state.}

\item{zeta}{The Initial Matrix for the Mixtures.

The initial matrix of the hyperparameters for the Dirichlet distribution used to describe
 the mixture probabilities at each location.}

\item{gamma_shape}{The Initial matrix for the Shape Parameters.

The initial matrix of the shape hyperparameters for each Gamma mixture component.}

\item{gamma_rate}{The Initial Matrix for the Rate Parameters.

The initial matrix of the rate hyperparameters for each Gamma mixture component.}

\item{obs}{The dataset.}

\item{mix.samples}{The indicator used to decide the data should come from different years.

The initial value is set to FALSE.}

\item{maxiter}{The maximum number of iterations the algorithm will do before stopping.

The algoritm will stop earlier than this number if the amount of improvement is smaller than a tolerance
 of 10^-9. The default value is 1000.}
}
\value{
A list of objects used to describe the model:

 * `priors`: A list containing all the starting matrices and probabilities.
 * `posteriors`: A list containing all the posterior matrices and probabilities.
 * `ELBO`: A vector tracking the ELBO as the model progresses (with spot `i` corresponding to iteration `i`).
 * `DIC`: A vector tracking the DIC as the model progresses (with spot `i` corresponding to iteration `i`).
 * `total_iters`: An integer counting the total number of interations done by the algorithm.
}
\description{
A function dedicated to the Variational Bayes EM process, simplified using Stochastic VI.
}
\details{
Used to run the Stochastic Variational Bayes EM Process for a Hidden Markov Model designed
 for rain data.
}
