% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBEM.gam.R
\name{VBEM.gam}
\alias{VBEM.gam}
\title{Variational Bayes EM Algorithm for Gamma Rainfall}
\usage{
VBEM.gam(
  D,
  S,
  Y,
  L,
  M = 2,
  xi,
  alpha,
  zeta,
  gammah,
  deltah,
  thetah,
  logbetah,
  obs,
  maxiter
)
}
\arguments{
\item{D}{Number of Days per Year.}

\item{S}{Number of States laid out in the model.}

\item{Y}{Number of Years collected.}

\item{L}{Number of locations data recorded in the data.}

\item{M}{Number of mixtures predetermined by user.

The initial number of mixtures is set to two, one for the zero rainfall cluster, the other
for the gamma distribution. Increasing the number of mixtures leads to instability.}

\item{xi}{The Current Matrix for the Initial States.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the initial probabilities each state.}

\item{alpha}{The Current Matrix for the Transitions.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the transition probabilities from state to state.}

\item{zeta}{The Current Matrix for the Mixtures.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the mixture probabilities at each location.}

\item{gammah}{The Current matrix for the gamma hyperparameters of the GC2 distribution.}

\item{deltah}{The Current Matrix for the delta hyperparameters of the GC2 distribution.}

\item{logbetah}{The Current Matrix for the log of the beta hyperparameters of the GC2 distribution.}

\item{obs}{The vector of observations.

The vector containing the y-values (typically the precipitation amounts) from the data.}

\item{maxiter}{The maximum number of iterations the algorithm will do before stopping.

The algorithm will stop earlier than this number if the amount of improvement is smaller than a tolerance
of 10^-9. The default value is 1000.}
}
\description{
A function dedicated to running the Variational Bayes EM algorithm.
}
\details{
Used to run the Variational Bayes EM Process for a Hidden Markov Model designed
for rain data. It is assumed the rainfall is calculated using a Gamma Distribution.
}
