% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VBM_function.R
\name{VBM.exp}
\alias{VBM.exp}
\title{Variational Bayes Maximization Formula}
\usage{
VBM.exp(
  numStates,
  numLoc,
  numMix,
  xi,
  alpha,
  zeta,
  gamma_shape,
  gamma_rate,
  q_1j,
  q_tj,
  q_tjml,
  q_jk,
  obs,
  iter
)
}
\arguments{
\item{numStates}{Number of States laid out in the model}

\item{numLoc}{Number of locations data recorded in the data}

\item{numMix}{Number of mixtures predetermined by user}

\item{xi}{The Current Matrix for the Initial States.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the initial probabilities each state.}

\item{alpha}{The Current Matrix for the Transitions.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the transition probabilities from state to state.}

\item{zeta}{The Current Matrix for the Mixtures.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
the mixture probabilities at each location.}

\item{gamma_shape}{The Current matrix for the Shape Parameters.

The current matrix of the shape hyperparameters for each Gamma mixture component.}

\item{gamma_rate}{The Current Matrix for the Rate Parameters.

The current matrix of the rate hyperparameters for each Gamma mixture component.}

\item{q_1j}{Posterior initial probability.

This matrix is usually the result of the VBE.exp function.}

\item{q_tj}{Posterior probability of stationary distribution.

This matrix is usually the result of the VBE.exp function.}

\item{q_tjml}{An augmented `q_tjml` array.

This matrix is usually the result of the VBE.exp function.}

\item{q_jk}{Posterior joint transition probability matrix.

This matrix is usually the result of the VBE.exp function.}

\item{obs}{The vector of observations.

The vector containing the y-values (typically the precipitation amounts) from the data.}

\item{iter}{The current iteration step.}
}
\value{
A list of objects built from empty objects in `var`:
 * `gamma_jml`: Posterior shape of exponential rate.
 * `delta_jml`: Posterior rate of exponential rate.
 * `xi_j`: Posterior Dirichlet paramters for initial distribution.
 * `alpha_j`: Posterior Dirichlet parameters for transition matrix rows.
 * `zeta_jl`: Posterior Dirichlet parameters for mixing probabilities.
 * `h_jml`: A matrix of coefficients for normalizing the mixture distributions.

 @export
}
\description{
A function dedicated to the maximization step in a Variational Bayes EM algorithm
}
\details{
Used as part of the Variational Bayes EM Process for a Hidden Markov Model designed
for rain data.
}
