% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stochastic_Max_function.R
\name{VBMS.exp}
\alias{VBMS.exp}
\title{Stochastic Maximization Formula}
\usage{
VBMS.exp(
  numStates,
  numLoc,
  numMix,
  numYears,
  numDays,
  xi,
  alpha,
  zeta,
  gamma_shape,
  gamma_rate,
  q_1j,
  q_tj,
  q_tjml,
  q_jk,
  obs,
  iter,
  priors
)
}
\arguments{
\item{numStates}{Number of States laid out in the model.}

\item{numLoc}{Number of locations data recorded in the data.}

\item{numMix}{Number of mixtures predetermined by user.}

\item{numYears}{Number of Years collected.}

\item{numDays}{Number of Days per Year.}

\item{xi}{The Current Matrix for the Initial States.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
 the initial probabilities each state.}

\item{alpha}{The Current Matrix for the Transitions.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
 the transition probabilities from state to state.}

\item{zeta}{The Current Matrix for the Mixtures.

The current matrix of the hyperparameters for the Dirichlet distribution used to describe
 the mixture probabilities at each location.}

\item{gamma_shape}{The Current matrix for the Shape Parameters.

The current matrix of the shape hyperparameters for each Gamma mixture component.}

\item{gamma_rate}{The Current Matrix for the Rate Parameters.

The current matrix of the rate hyperparameters for each Gamma mixture component.}

\item{q_1j}{Expectation of latent variable s_1j.

The latent variable s_tj represents if, at time t, the state is j.
 s_1j is the initial states.}

\item{q_tj}{Expectation of latent variable s_tj.

The latent variable s_tj represents if, at time t, the state is j.}

\item{q_tjml}{Expectation of latent variable r_tjml.

The latent variable r_tjml represents an indicator variable with 1 if y_tl (the data point at time t and location l)
 comes from mixture m, and s_t = j.}

\item{q_jk}{Expectation of state transistions

The latent variables s_tj*s_t+1,k represent the movement from state j to k from time t to t+1.}

\item{obs}{The dataset.}

\item{iter}{The current step of the iteration process.}

\item{priors}{A list containing the priors used at the start of the process.

The priors are needed for xi, alpha, zeta, gamma_shape, and gamma_rate.}
}
\value{
A list of objects:

 * `gamma_jml`: Posterior shape of exponential rate.
 * `delta_jml`: Posterior rate of exponential rate.
 * `xi_j`: Posterior Dirichlet paramters for initial distribution.
 * `alpha_j`: Posterior Dirichlet parameters for transition matrix rows.
 * `zeta_jl`: Posterior Dirichlet parameters for mixing probabilities.
 * `h_jml`: A matrix of coefficients for normalizing the mixture distributions.
}
\description{
A function dedicated to the maximization step in a Variational Bayes EM algorithm, simplified using Stochastic VI.
}
\details{
Used as part of the Stochastic Variational Bayes EM Process for a Hidden Markov Model designed
 for rain data.

Stochastic VI assumes exchangeablility for the emission distributions used for each year. Thus
 rather than trying to compute N data points, we split the data with N = D*Y, days and years, then account for the number of years
 by multiplying parts of the update by Y. However, the user must decide some step size tau. The step size must
 fulfill the Robbins-Monro conditions.

 In particular, this function uses stochastic gradient ascent with step size tau = 1/iter.
}
