% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi_function.R
\name{viterbi.decoding}
\alias{viterbi.decoding}
\title{Viterbi Decoding Function}
\usage{
viterbi.decoding(b, a, initDist, numDays, numYears, numStates)
}
\description{
Left here as a note to fix later.
}
