% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi_function.R
\name{viterbi.decoding}
\alias{viterbi.decoding}
\title{Viterbi Decoding Function}
\usage{
viterbi.decoding(b, a, initDist, numDays, numYears, numStates)
}
\arguments{
\item{b}{Matrix of Emission Probabilities}

\item{a}{Matrix of Transition Probabilities}

\item{initDist}{The Vector of Initial Probabilities for each state.}

\item{numDays}{Number of Days in the Data}

\item{numYears}{Number of Years recorded}

\item{numStates}{Number of States decided}
}
\value{
A vector that represents the best state sequence.
}
\description{
The Viterbi Algorithm applied to Rain Models.
}
